+++ example/index.ts
@@ -1,17 +1,116 @@
-import { createMapper } from '../src'
+import {
+  field,
+  fieldArray,
+  model
+} from '../src'
 
-setTimeout(() => {
+const ServerData = {
+  DeepInfo: {
+    BadHabits: ['Coffee', 'Development']
+  },
+  Family: {
+    ChildrenCount: 0,
+    Spouse: false,
+  },
+  FirstName: 'Sergey',
+  ID: '1CASSD@D#@Dd2d@2dDFC',
+  Job: {
+    Exp: 4,
+    Skills: ['JavaScript', 'ReactJS', 'NodeJS']
+  },
+  Languages: [
+    {
+      ID: '1',
+      Name: 'English',
+    },
+    {
+      ID: '2',
+      Name: 'Russian',
+    },
+  ],
+  LastName: 'Volkov',
+}
 
-  const mapy = createMapper({
-    firstName: 'FirstName',
-    lastName: 'LastName',
+class YourModelDeclaration {
+  clientProperty = fieldArray('Original_Property', 'string')
+}
+
+const YourModel = model(YourModelDeclaration)
+
+console.log('YourModel', YourModel)
+
+class ProfileDeclaration {
+  badHabits = field(
+      ({ DeepInfo }: any) => DeepInfo.BadHabits,
+      ({ badHabits }) => ({ DeepInfo: { BadHabits: badHabits } })
+    )
+  family = field('Family', {
+    childCount: field('ChildrenCount', 'number', 'string'),
+    spouse: field('Spouse', 'boolean')
   })
+  id = field('ID')
+  job = field('Job', {
+    experience: field('Exp', 'integer'),
+    skills: fieldArray('Skills', 'string')
+  })
+  languages = fieldArray('Languages', {
+    id: field('ID', 'string'),
+    name: field('Name', 'string')
+  })
+  personalInfo = field(
+      ({ FirstName, LastName }) => ({
+        firstName: FirstName,
+        fullName: `${FirstName} ${LastName}`,
+        lastName: LastName,
+      }),
+      ({ personalInfo }) => ({
+        FirstName: personalInfo.firstName,
+        LastName: personalInfo.lastName,
+      })
+    )
 
-  const cc = mapy.toClient({
-    FirstName: 'Sergey',
-    LastName: 'Volkov',
-  }) as any
+  getFullName = () => (this.personalInfo as any).firstName
+}
 
-  console.log('cc.firstName', cc.firstName)
+const ProfileModel = model(ProfileDeclaration)
 
-}, 14000)
+const profile = new ProfileModel(ServerData)
+
+profile.personalInfo.firstName = 'Sergey'
+profile.personalInfo.lastName = 'Volkov'
+
+profile.id = `${profile.id}_CHANGED`
+
+console.log(profile.deserialize())
+
+const SomeModel = model({
+  prop1: field('Prop')
+})
+
+console.log(new SomeModel({ Prop: 'blabla ' }))
+
+class AnimalD {
+  age = field('Age', 'number')
+  name = field('Name', 'string')
+}
+
+class DogD extends AnimalD {
+  breed = field('Breed', 'string')
+}
+
+const DogModel = model(DogD)
+
+const dog = new DogModel({
+  Age: 4,
+  Breed: 'scottish terrier',
+  Name: 'Fluffy'
+})
+
+console.log(dog)
+  /*
+  {
+    age: 4,
+    breed: 'scottish terrier',
+    name: 'Fluffy'
+  }
+  */