+++ src/constants.ts
@@ -1,0 +1,12 @@
+
+export const TYPE_OF_CLASS_PROP_VALUE = '@@VALUE_TYPE'
+export const NAME_OF_CLASS_PROP = '@@CLASS_PROP_NAME'
+export const DECLARATION_PROP = '@@property_declaration'
+
+export enum SchemeType {
+  ONE_STRING = '@ONLY_STRINGS',
+  TWO_STRINGS = '@ONLY_STRINGS',
+  THREE_STRINGS = '@ONLY_STRINGS',
+  STRING_AND_CLASS = '@STRING_AND_CLASS',
+  SERIALIZERS = '@SERIALIZERS',
+}