+++ src/global_types.ts
@@ -1,0 +1,5 @@
+export type ValueOf<T> = T[keyof T]
+
+export interface AllKeysAre<T> {
+  [key: string]: T
+}