+++ CHANGELOG.md
@@ -3,10 +3,16 @@
 <!-- TEMPLATE OF NEW VERSION -->
 
 <!-- 
-## [VERSION](https://github.com/acacode/mapy/releases/tag/VERSION)
+## [VERSION](https://github.com/acacode/serializy/releases/tag/VERSION)
 
 ### Changed
 ### Fixed
 ### Added
 ### Removed
  -->
+
+## [0.0.1-alpha](https://github.com/acacode/serializy/releases/tag/0.0.1-alpha)
+
+### Added
+- Created a library (rollup build, small documentation)  
+- Added functions: `field()`, `fieldArray()`, `model()`